+++ .gitignore
@@ -69,3 +69,4 @@ local.properties
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
+/lib/
