+++ jmulticard-jse/src/test/java/es/gob/jmulticard/TestAccv.java
@@ -11,6 +11,7 @@ import es.gob.jmulticard.card.gide.smartcafe.SmartCafePkcs15Applet;
 import es.gob.jmulticard.card.iso7816four.Iso7816FourCard;
 import es.gob.jmulticard.jse.provider.CachePasswordCallback;
 import es.gob.jmulticard.jse.provider.ProviderUtil;
+import es.gob.jmulticard.jsr268.AOUtil;
 
 /** Pruebas de las tarjetas G&amp;D de ACCV.
  * @author Tom&aacute;s Garc&iacute;a-Mer&aacute;s. */
